Updating project layout and builds (#21)
Updating the layout of the project to facilitate the seamless addition of new tasks in the future. This change also includes updates to the build pipelines to perform greater validations, as well as increased use of GitHub features.

This change furthermore means that the GitHub Community Profile will reach 100% after merging to main.