PRMetrics: Enhancing testing (#71)
Reinstating additional testing that was previously removed to address an urgent functionality break. This reinstatement also includes some logic changes that were found to be necessary following the reinstatement of the tests, specifically around checks as to whether comment content exists or not.

This change also includes some settings updates related to NPM packages and ESLint.