Upgrading agent version to v2.186.1 (#83)
Upgrading the agent version to the latest release, v2.186.1. Azure DevOps servers without this release will automatically pull down the update, so this will not reduce the number of servers able to run the extension.

This change also includes an update to the extension casing, for enhanced readability within the server diagnostic messages.