M✔ ◾ PR Metrics v1.3.0: Full GitHub support (#130)
Updating PR Metrics to v1.3.0 to add full GitHub support. This expanded GitHub support now adds comments for the metric statistics and files that do not need to be reviewed. It supports GitHub, GitHub Enterprise, and GitHub AE. The documentation has also been updated to reflect this.

This change involved refactoring some existing method logic to use a generic interface that would support both Azure DevOps and GitHub.

This change also includes a bug fix to better facilitate running pipelines on multiple platforms with different line endings.

It should be noted that some of the functionality differs between Azure DevOps and GitHub due to the different commenting models on the two platforms. However, this should have negligible effect on the experience.