XS✔ ◾ PR Metrics v1.3.1: Handling large files on GitHub (#168)
The GitHub APIs changed to prevent the addition of comments to files whose full diff cannot be displayed. This resulted in exceptions being thrown in certain cases, but the correct functionality would be to skip the addition of comments to these files.

This PR catches the exception, then logs and suppresses it. Unfortunately, there appears to be no one to tell in advance whether a comment creation will fail, so we cannot avoid the error.

It also updates the dependencies, licensing information, and release version to facilitate a new deployment.