Downgrading Minimum Agent Version (#183)
Azure DevOps Server (i.e., the non-cloud version of Azure DevOps) comes pre-packaged with agent version 2.181.2 and not the latest verson used within the current task. This means that the agent continually needs to be updated, which results in poor runtime performance that is not in keeping with the design goals of the task.