XS✔ ◾ v1.4.2: Bug fix for GitHub API restrictions (#221)
The GitHub API will fail to add a comment to a file in a PR if the file diff is larger than a specific threshold. In these cases, PR Metrics will throw an error, preventing further execution.

The error is now caught and handled in the cases where it is expected, but still thrown for other errors. There has also been some refactoring to extract out shared logic. Tests have been updated.

In addition, the version of PR Metrics has been updated to v1.4.2 to allow this change to be released to the marketplaces.