S✔ ◾ Updating @octokit/plugin-rest-endpoint-methods to latest (#303)
There was a dependency on @octokit/plugin-rest-endpoint-methods version 6.2.0. This introduced risk as the dependency could develop security vulnerabilities over time, and it would not be possible to migrate to the latest version to resolve these.

This dependency was updated to the latest version by changing the API calls to remove the now unsupported dependency on the GitHub diff position and to replace it with a diff line number. This involves reading the diff and parsing it to find the appropriate line number. New logic and dependencies have been introduced for this.

This change also includes:
- Some refactoring to support the above changes.
- Updating of all dependencies to the latest versions.
- Refactoring of integration tests into unit tests, as the integration tests stopped working due to a changed dependency that does not support test environments.
- Refactoring of the product code to enable the integration tests to be run as unit tests, due to the changes outlined above.
- Removal of superfluous test logic following the test refactoring.
- Onboarding of Babel to rewrite some dependencies to an older version of Node.js (v10), in order to continue supporting Azure DevOps. This rewriting is not performed for GitHub, where the version of Node.js used is much more recent.