XS⚠️ ◾ Removing support for System.AccessToken in favour of PR_Metrics_Access_Token (#305)
Removing support for System.AccessToken in favour of the more compliant PR_Metrics_Access_Token.

System.AccessToken was temporarily allowed with a warning, but it is now being removed.

Removing the items highlighted via the TODO comments. The code for failing with an error when PR_Metrics_Access_Token is missing is already present.