XS⚠️ ◾ GitHub Workflow Enhancements (#349)
It is currently time consuming to release an updated version of PR Metrics due to the number of places requiring a version number update and the steps involved in preparing the .vsix file. This PR is designed to simplify the process by automating as much as possible and providing detailed instructions where manual tasks are still required.

Automating tasks related to project maintenance:
- Creating workflows for automatically updating the version number and creating a new release.
- Enhancing workflows to reduce Dependabot PR effort.
- Enhancing workflows to reduce access level of Personal Access Token (PAT).
- Adding GitHub Actions Dependabot updates, as updates for these will otherwise not occur.
- Improving documentation related to the above points.
- Adding information about handling GitHub fine-grained PATs and workflow permission limiting.
This PR also updates the dependencies and related files to the latest versions.