XS✔ ◾ Adding AlwaysCloseComment input (#350)
Fixes #256 by allowing for the size and test comment to be optionally automatically closed.

Adding a new input that is only supported by Azure DevOps (as the logic doesn't make sense in the context of GitHub due to its differing behaviour). This will automatically resolve the size and test comment to prevent it blocking review closure.

This PR also includes some minor pipeline fixes, for issues that were detected while running this PR through the updated pipelines.