XS✔ ◾ Ignore whitespace when calculating metrics (#357)
Fixes #356 by ignoring whitespace-only lines returned when calculating the PR Metrics.

Consideration was given to making this an optional setting, but it was decided that this behaviour is preferable in all cases as it will more closely align with customer expectations. Moreover, it aligns more closely with Azure DevOps behavior. It will be released as part of a minor (rather than patch) release.