Pipeline permissions refinement (#365)
To improve the security score of the repo within the OSSF Scorecard, the pipeline permissions are being updated.

For the pipelines and jobs where no permissions are required, the permissions are being set to read-all. This change effectively makes no difference to the current security posture, but it allows the automated tooling to determine that best practices are being used.