XS✔ ◾ Enhancing Glob Support (#373)
Multiple negative globs are processed in an unexpected way, whereby only the last negative glob is considered when determining whether to exclude a file. This is in response to #372.

The match determination logic failed to consider the state from previous negative matches. The logic has now been adapted to consider this an ensure an exclusion on one line is not overridden by an exclusion on a subsequent line.

The changes also include support for double exclusions, in line with Azure DevOps glob support.