XS✔ ◾ Validating PR IDs (#393)
## Summary

### Motivation

The Pull Request ID is required for calling Azure DevOps and GitHub
APIs.

For GitHub, this is coming from the variable `GITHUB_REF`. But the
variable is only validly set when using `pull_request` as the build
trigger according to
<https://docs.github.com/actions/learn-github-actions/variables>.

It is beneficial as per #382 to better flag this error and provide
guidance as to how to remediate the issue.

### Technical

Updating the code to fail gracefully when the Pull Request ID is
unavailable. The message will provide instructions as to the correct
course of action to take.

Additional documentation is being added to the `README` to provide
further assistance.

Note that, unfortunately, it is not possible to support the
`pull_request_target` trigger as the Pull Request ID is unavailable when
using this.

The update also updates all dependencies.

Resolves #382.

## Testing

### Test Types

- [X] Unit tests
- [X] Manual tests

### Unit Test Coverage

100%