XS✔ ◾ Updating Release Pipeline (#496)
## Summary

Relocating the Visual Studio Marketplace publishing logic from GitHub to
Azure DevOps, to switch from PATs to Workload Identity Federation. This
update also includes a small optimisation when testing the latest PR
Metrics release.

## Detailed Copilot Generated Summary (manually refined)

These changes include significant updates to the Azure DevOps pipeline
configuration files and the GitHub workflow file. They focus on
modifying the release process and improving the build process. The most
significant changes are the addition of a new `release.yml` file, the
modification of the `template.yml` file to change the build stages, and
the removal of the "Publish Release" step from the `release-phase-2.yml`
file.

Addition of new configuration:

*
[`.github/azure-devops/release.yml`](diffhunk://#diff-25f998e817515523e95edd3b4e0eb06fad5909deec7e3d4b7d57f4912cb39349R1-R110):
A new configuration file has been added that sets up a new Azure DevOps
pipeline for releases. The pipeline is triggered by tags starting with
'v' and does not run for pull requests. It includes jobs for checks,
release, and publishing to the Azure DevOps marketplace.

Modifications to existing configuration:

*
[`.github/azure-devops/template.yml`](diffhunk://#diff-9ba73744bad564cb8afb7eac4a1ffce0822431667e90f80de88300f094aae565L15-R30):
The build stages have been modified. The "Deploy" job has been replaced
with a "Prerequisites" job. The "Delay" job has been removed and the
"PRMetrics_macOS" job now depends on the "Prerequisites" job. These
changes streamline the build process and optimize the use of resources.
[[1]](diffhunk://#diff-9ba73744bad564cb8afb7eac4a1ffce0822431667e90f80de88300f094aae565L15-R30)
[[2]](diffhunk://#diff-9ba73744bad564cb8afb7eac4a1ffce0822431667e90f80de88300f094aae565L72-R82)

Removal of redundant configuration:

*
[`.github/workflows/release-phase-2.yml`](diffhunk://#diff-c7e4bd8feb36835f6722b91d36ca67dcad2ece6f364e0404754016195651f1b0L72-L79):
The "Publish Release" step has been removed from the workflow. This step
was previously used to publish the extension to the Azure DevOps
marketplace. The new `release.yml` file now manages this process.