XS✔ ◾ Updating CodeQL configuration (#498)
## Summary

This change primarily improves the clarity of the pull request template
and reorganizes the CodeQL validation workflow, moving this from the
`build.yml` workflow to a new `codeql.yml` workflow.

Simplification of Pull Request Template:
*
[`.github/pull_request_template.md`](diffhunk://#diff-b2496e80299b8c3150b1944450bd81c622e04e13d15c411d291db0927d75fd6bL3-R5):
The pull request template was simplified by removing the separate
"Motivation" and "Technical" sections and replacing them with a single
comment instructing users to provide a summary of the reasons or
motivations behind the PR, what it aims to accomplish, and technical
implementation details.

Reorganization of CodeQL Validation Workflow:
*
[`.github/workflows/build.yml`](diffhunk://#diff-5c3fa597431eda03ac3339ae6bf7f05e1a50d6fc7333679ec38e21b337cb6721L273-L290):
The "validate-codeql" job was removed from this workflow.
*
[`.github/workflows/codeql.yml`](diffhunk://#diff-12783128521e452af0cfac94b99b8d250413c516ec71fe6d97dbea666ff7ba27R1-R45):
A new workflow file was created for the CodeQL validation job. This job
is set to run on push and pull request events for the main branch, on a
weekly schedule, and on workflow dispatch events. The job includes steps
to checkout the code, initialize CodeQL, and analyze the code.