XS✔ ◾ ADO Pipelines: Secure Variables Change – Part 2 (#503)
## Summary

This complements #502 by making the same changes to the Production
pipeline, to ensure access to the relevant secure variables.

It comprises a minor change to the `.github/azure-devops/prod.yml` file.
A new variable group `PR Metrics` has been added to the variables
section. This change will allow the pipeline to access shared variables
defined in the `PR Metrics` group.