M⚠️ ◾ Modernising Linting (#511)
## Summary

Updating ESLint to use the modern configuration file and work with the
newer releases. This resulted in some new violations, many of which have
been fixed as part of these changes. Others are suppressed and will be
considered for resolution in the future.

### Detailed Description

This includes several changes mainly aimed at improving code quality,
updating the version of the software, and modifying the test coverage
requirements. The most significant changes include the modification of
the `.c8rc.json` file to require 100% test coverage, removal of the
`.eslintrc.yml` file, addition of the `eslint.config.mjs` file, updates
to the version of the software in several files, and various code
quality improvements in TypeScript files.

**Test coverage requirement changes:**

*
[`.c8rc.json`](diffhunk://#diff-9fd2b089a79150b0fd1cad8f8961358547e3cc3256f8466c8602c647e5734de1L7-R10):
Modified the test coverage requirements to 100% for statements,
branches, functions, and lines.

**ESLint configuration changes:**

*
[`.eslintrc.yml`](diffhunk://#diff-9e1ecc14c733bb1ae2e523089f1262ac6ffccbcf950487ee0984403603550e57L1-L18):
Removed this file.
*
[`eslint.config.mjs`](diffhunk://#diff-9601a8f6c734c2001be34a2361f76946d19a39a709b5e8c624a2a5a0aade05f2R1-R82):
Added this file to configure ESLint.

**Version updates:**

*
[`.github/workflows/release-phase-1.yml`](diffhunk://#diff-a775c174b4973a7a42b20039065ebe3444ca9b3e3d28ee17ec7f8da67924460bL24-R25):
Updated the minor version and reset the patch version.
*
[`.github/workflows/support/release-trigger.txt`](diffhunk://#diff-4a1ff8fb39a32cd5133d791b3ed91ed72e8ab200b9a45f157f2c422a94a5c11fL1-R1):
Updated the version.
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L116-R116):
Updated the version in two places.
[[1]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L116-R116)
[[2]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L126-R126)
*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L5-R20):
Updated the version.
*
[`src/task/Strings/resources.resjson/en-US/resources.resjson`](diffhunk://#diff-c0fc6893012d5b83ce394e09ff298a469483e5eec91edd15b9744fc81db6f598L5-R5):
Updated the version.

**Code quality improvements:**

*
[`src/task/index.ts`](diffhunk://#diff-445f33c5199a2e71fde9062a5f8c6b5237e3230e83e8be8fcfbf06de48abd5abL1-R12):
Modified the run function to be an arrow function and added await to the
call to `pullRequestMetrics.run(__dirname)`.
*
[`src/task/src/git/gitInvoker.ts`](diffhunk://#diff-393c3008fe54c65d56363117b91baeb2bfacc051055e6fbffc71df7ea919cc28L1-R10):
Made several changes to improve code quality, such as using radix
parameter in `parseInt` calls, removing unnecessary else blocks, and
changing the `for` loop to use `+= 1` instead of `++`.
[[1]](diffhunk://#diff-393c3008fe54c65d56363117b91baeb2bfacc051055e6fbffc71df7ea919cc28L1-R10)
[[2]](diffhunk://#diff-393c3008fe54c65d56363117b91baeb2bfacc051055e6fbffc71df7ea919cc28L56-R58)
[[3]](diffhunk://#diff-393c3008fe54c65d56363117b91baeb2bfacc051055e6fbffc71df7ea919cc28L87-R89)
[[4]](diffhunk://#diff-393c3008fe54c65d56363117b91baeb2bfacc051055e6fbffc71df7ea919cc28L160-R163)
[[5]](diffhunk://#diff-393c3008fe54c65d56363117b91baeb2bfacc051055e6fbffc71df7ea919cc28L169)
[[6]](diffhunk://#diff-393c3008fe54c65d56363117b91baeb2bfacc051055e6fbffc71df7ea919cc28L183-R187)
*
[`src/task/src/git/octokitGitDiffParser.ts`](diffhunk://#diff-0e959b5630206a4ce8a405c21439f051d4bd44fee10de7c1976ba982365899bbL69-R79):
Changed the `diffResponses` split to use the `u` flag in the regular
expression and modified the `for` loop to use template literals.
*
[`src/task/src/metrics/codeMetrics.ts`](diffhunk://#diff-f34569608b714fbc5a7b215183b770a19aec3c8056562c96219e2bcaea0fb794L45-R47):
Made the constructor public and changed an `Error` throw to use `new
Error`.
[[1]](diffhunk://#diff-f34569608b714fbc5a7b215183b770a19aec3c8056562c96219e2bcaea0fb794L45-R47)
[[2]](diffhunk://#diff-f34569608b714fbc5a7b215183b770a19aec3c8056562c96219e2bcaea0fb794L138-R140)

## Testing

### Test Types

- [X] Unit tests
- [X] Manual tests

### Unit Test Coverage

100%

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>