XS⚠️ ◾ Declaring method static (#570)
## Summary

This changes the `invokeApiCall` method in several classes to make it a
static method in the `BaseReposInvoker` class and update its usage
accordingly. The most important changes include modifying the
`invokeApiCall` method in `AzureReposInvoker`, `BaseReposInvoker`, and
`GitHubReposInvoker` classes.

Changes to `invokeApiCall` method:

*
[`src/task/src/repos/baseReposInvoker.ts`](diffhunk://#diff-a221826b34914e756e969871b529aa4d4428c7429ce15f23e537278dd6186c31L26-R26):
Changed the `invokeApiCall` method to be a static method.
*
[`src/task/src/repos/azureReposInvoker.ts`](diffhunk://#diff-dd04424eb7f605d4b4584e72cc84c561663aeaad32743966ed30e904a38959e0L329-R329):
Updated the `invokeApiCall` method call to use the static method from
`BaseReposInvoker`.
*
[`src/task/src/repos/gitHubReposInvoker.ts`](diffhunk://#diff-f940089e58c7285d97fbb04b132d2827a8aaf024094dfcd7ba27351b7ae6b4a3L260-R260):
Updated the `invokeApiCall` method call to use the static method from
`BaseReposInvoker`.

## Testing

### Test Types

- [X] Unit tests
- [ ] Manual tests

### Unit Test Coverage

100%

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>