XS✔ ◾ GitHub Actions: Correct Permissions for Release (#573)
## Summary

This includes a change to the GitHub Actions workflow configuration in
the `.github/workflows/release-phase-2.yml` file. The change updates the
permissions for the `release` job to allow writing to the repository
contents, which will fix issues encountered during the release process.

Permissions update:

*
[`.github/workflows/release-phase-2.yml`](diffhunk://#diff-c7e4bd8feb36835f6722b91d36ca67dcad2ece6f364e0404754016195651f1b0L21-R22):
Modified the `permissions` section for the `release` job to grant write
access to repository contents.