XS⚠️ ◾ Updating external URLs (#604)
## Summary

This includes updates to configuration files, documentation, and source
code to improve link accuracy and maintainability. The most important
changes involve updating outdated URLs to their correct locations and
enhancing configuration for link checking.

### Link updates for accuracy:

*
[`.github/linters/markdown-link-check.json`](diffhunk://#diff-7fe04f8c8647e157cfa10cde0746b50a5e4ab1f4cfee31c63fa8bbaac2b716fcR2):
Added `429` to the `aliveStatusCodes` to account for rate-limited
responses as valid during link checks.
*
[`docs/default-code-file-extensions.md`](diffhunk://#diff-6f62219a8fd05e168e5acb123bf2d15ff9ea805bb079b250069d0b501b51bb50L225-R225):
Updated the link to GitHub Linguist's language definitions from the
outdated `master` branch to the current `main` branch.
*
[`src/task/src/metrics/inputsDefault.ts`](diffhunk://#diff-14f0e4ce4ae1fa354e284be06c26cecfc06af0b86badaedbaab279192c01eefeL45-R45):
Updated the reference to GitHub Linguist's language definitions in the
comments from the `master` branch to the `main` branch.