XS✔ ◾ Correcting Release Pipeline (#606)
## Summary

This updates the release pipeline configuration to refine the condition
for terminating builds. The change ensures that builds are not
terminated for branches starting with 'v'.

Pipeline configuration changes:

*
[`.github/azure-devops/release.yml`](diffhunk://#diff-25f998e817515523e95edd3b4e0eb06fad5909deec7e3d4b7d57f4912cb39349L56-R56):
Updated the `condition` in the "Terminate on PR" step to terminate
builds unless the branch name starts with 'v', instead of only allowing
builds on the 'main' branch.