XS✔ ◾ Release v1.7.5: Licenses update (#640)
## Summary

This pull request updates the GitHub Actions workflows for the release
process to improve security and consistency. The most significant
changes include standardizing the use of a fine-grained personal access
token (PAT) and refining permissions for better access control.

### Updates to token usage:
*
[`.github/workflows/release-phase-1.yml`](diffhunk://#diff-a775c174b4973a7a42b20039065ebe3444ca9b3e3d28ee17ec7f8da67924460bL33-R33):
Updated the token used for checkout to `secrets.RELEASE_CHECKOUT` for
consistency across workflows.
*
[`.github/workflows/release-phase-2.yml`](diffhunk://#diff-c7e4bd8feb36835f6722b91d36ca67dcad2ece6f364e0404754016195651f1b0R32-R43):
Added a `token` parameter to the checkout step, using
`secrets.RELEASE_CHECKOUT` with detailed permissions for better
security.

### Refinements to permissions:
*
[`.github/workflows/release-phase-2.yml`](diffhunk://#diff-c7e4bd8feb36835f6722b91d36ca67dcad2ece6f364e0404754016195651f1b0L22-R22):
Changed permissions from `contents: write` to `pull-requests: write` to
align with the specific requirements of the workflow.

### Minor adjustments:
*
[`.github/workflows/release-phase-2.yml`](diffhunk://#diff-c7e4bd8feb36835f6722b91d36ca67dcad2ece6f364e0404754016195651f1b0R32-R43):
Removed the "v" prefix when writing the version to the GitHub output in
the `Get Version` step for consistency.

## Testing

### Test Types

- [X] Unit tests
- [X] Manual tests

### Unit Test Coverage

100%